// src/app/shared/interfaces/bike.interface.ts
export interface IBike {
    nameEN: string;
    type:string;
    price: number;
    count: number;
    weight: number;
    size: string;
    color: string;
    image: string;
    fork: string;
    steeringColumn: string;
    shifters: string;
    connectingRodSystem: string;
    brake: string;
    frontSwitch: string;
    rearSwitch: string;
    cassette: string;
    shain: string;
    wheels: string;
    col: string;
    reinforcingPin: string;
    collar: string;
    category: string;
    id?:string,
    buyCount?:number,
    data?:string
}